Code in Jupyter Notebook
Basic conclusions:

Taking into account the population, including people who tested positive for COVID who died, the most deaths were recorded with the simultaneous diagnosis of:

- pneumonia;
- high blood pressure;
- diabetes.

In the case of pneumonia, as a co-morbidity with COVID, mortality is very high for practically every age group and amounts to 75%. The only exception is the 0-18 age group, where the mortality rate is 67%.

In the case of hypertension, as a co-morbidity with COVID, mortality for virtually every age group is age-dependent and increases exponentially. The highest mortality rate concerns the age groups 71 - 80 and 80 and above - it amounts to 55% and 54%, respectively. The large increase in mortality is particularly noticeable for people over 40 and over 18 years of age. In the two oldest age groups, the mortality rate exceeds 50% (age group 71-80 and over 80). The best results (95% of patients survived) were recorded in the age group up to 18 years. This is probably related to the characteristics of the disease, which mainly affects the elderly.

In the case of diabetes, as a co-morbidity with COVID, the highest mortality rate concerns people between 51 and 80 years of age. Importantly, in a rare age group, the percentage of mortality does not exceed 50% - the highest value applies to the age group of 61 - 70 and amounts to 44%. As in the case of hypertension, the highest survival rate is in the age group 0 - 18 years (96% of patients survived). This is probably related to the characteristics of the disease, which mainly affects the elderly.

Conclusions: the highest risk of death in the case of covid was diagnosed for people over 50 years of age, with comorbidities (pneumonia, hypertension, diabetes).

Recommendations:

- vaccinations for people from the risk group (in the absence of medical contraindications);
- isolation of people at high mortality risk from potentially infected people (e.g. in public places);
- increasing funds for the treatment of people with diseases from the TOP-3 list of comorbidities;
- education and prevention in the case of hypertension and diabetes (promoting a healthy lifestyle).